board

+++ carbonlease/src/main/java/com/kh/board/model/dto/BoardDTO.java
@@ -23,7 +23,7 @@ public class BoardDTO {
 	private int viewCount;
 	private Date enrollDate;
 	private char status;
-	private int memberNo;
+	private long memberNo;
 	private int regionNo;
 	private String regionName;
 	private String nickname;
